update: collect all the extra functions together
